clients/python/requirements.txt
aiohttp==3.8.1
gql==3.0.0
python-dotenv==0.19.2
